Reject Web Serial requests with an opaque origin
The Web Serial API tracks permissions using the origin of the top-level document in the frame tree. If this document has an opaque origin then there is no way to format the origin for display to the user in permission prompts or to write their decision in the preferences file.

Access to the Web Serial API from such contexts should therefore be blocked.

Ignore opaque origins request to connect to serial devices

Bug: 1375133
Change-Id: I4552ae74d480aa8df9ff93527fc85618bc03b947