Fix storage bucket quota overflow
This fixes overflow and underflow issues with storage buckets. Creating
a bucket with a quota 2^31 or greater would cause the bucket's quota to
overflow when stored in the quota database. A negative number would
overflow the other way when processed by v8.

This updates all variables representing quota to signed int64s to match
how it is represented in the database. It also adds the EnforceRange
attribute to the idl, so that it automatically errors when out of range
(-(2^53) to 2^53-1).

This also adds checks for negative quotas and errors if there is one.

Bug: 1406355
Change-Id: I5ba8d7716e8ec9735b2da453ff3e6a90e47211e4