Move renderer-side same-url-replacement enforcement to LocalFrame::NavigationShouldReplaceCurrentHistoryEntry
When we navigate to the current url, we treat it as a replacing
navigation, rather than pushing a new history entry. The browser
process enforces this most of the time, but there's a special case
to enforce it in the renderer for synchronous same-document commits.
This renderer process special case is in a weird place late in the
commit process.

The late timing is also a problem because the navigation API wants
to report whether the navigation will replace the current history
entry when it fires the navigate event. The navigate event fires
before we enforce this policy currently.

By moving same-url replacement to
LocalFrame::NavigationShouldReplaceCurrentHistoryEntry,
we duplicate some logic between the renderer and the browser
process, but we consistently switch the navigation to be replacing
before the navigate event fires, ensure we expose the correct
navigation type to JS.

Change-Id: I8859ac19fddbfc2f9aa42cad8e19b537df9679b1