navigation.navigate(url, { history: "push" }); should actually force push
Currently, { history: "push" } is actually push-or-fail: if the
navigation would've been a push, it pushes. If it would've been a
replace, the navigation fails immediately.

This CL makes { history: "push" } actually override some cases where
we would implicitly convert a push navigation to a replace:
navigations before onload completes, and navigation to the current url.

Change-Id: I0b985a4c9563c6000f7ee44153394466cc4bcb50