view-transitions: Use GeometryMapper for computing transforms.
GeometryMapper is the better source of truth to compute the screen space
transform for a DOM element. Use that instead of the algorithm which
relies on state computed during layout.

This avoids cases where we end up with a fractional transform on the
element while the fractional component is handled as a part of
rasterizing the element.

Bug:1405743
Change-Id: Ic35cf5c81353bdca07c203a1745df93d89db2524