[mathml] Fix offset of vertical glyph assembly
When ShapeResult::CreateForStretchyMathOperator shapes vertical glyph
assemblies, the offset of each part is adjusted by the advance specified
in the corresponding MathGlyphVariantRecord table [1]. Note that this may be slightly different from the calculated height of the glyph, but
is faster to get. In any case the baselines of glyphs are used as a
reference when setting an offset in ShapeResults, so this CL changes the
adjustment to use the glyph ascent instead of the glyph height.

This CL fixes rendering issues for vertical stretchy MathML Operators
for fonts like Cambria Math that use non-zero descent for parts in a
glyph assembly.

StretchyOperatorShaperTest.GlyphVariants is tweaked to use an equivalent
stretchy.woff font (glyphs have zero descent) and a new test
StretchyOperatorShaperTest.GlyphVariantsCenteredOnBaseline is introduced
for a similar stretchy-centered-on-baseline.woff (glyphs have non-zero
and equal ascent/descent).

Finally, a WPT test is added to check that the painting of a vertical
glyph assembly with the two fonts above matches the location of their
bounding boxes.

[1] https://learn.microsoft.com/en-us/typography/opentype/spec/math

Bug: 1409380
Change-Id: I173c1cf461cebe3523e1a10aaf8d7b479bca03ad