Fix WPT testcases that do not properly account for counter-style fallback behavior of prefix/suffix descriptors.
Per spec, "If a counter style is used to represent a counter value outside of its ranges,
the counter style instead drops down to its fallback counter style."

  https://www.w3.org/TR/css-counter-styles-3/#counter-style-range
  https://www.w3.org/TR/css-counter-styles-3/#counter-style-fallback

Some existing tests incorrectly assumed that a prefix or suffix from a custom rule will still apply
even when the counter value is not in range. That would mean "fallback" is only partial, using
the symbols/system from the fallback style mixed with the prefix/suffix of the originally-specified
style, which prevents use-cases like that in bug 1808995 working as expected.

Differential Revision: https://phabricator.services.mozilla.com/D167410

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1808995
gecko-commit: c2a49ee389d95fef709a326f089033aa362543aa
gecko-reviewers: emilio